🏷️ Take a u32 radius for circle and circle_fill
circle/circle_fill (and circ/circfill) now take i32 centres and a u32 radius,
staying infallible: a zero radius still draws a single pixel, which the
rasterizer already handles, so there is no silent no-op to surface.

+++ rico8/src/ffi.rs
@@ -23,8 +23,8 @@ extern "C" {
     pub fn line(x0: i32, y0: i32, x1: i32, y1: i32, color: i32);
     pub fn rect(x0: f32, y0: f32, x1: f32, y1: f32, color: i32);
     pub fn rect_fill(x0: f32, y0: f32, x1: f32, y1: f32, color: i32);
-    pub fn circle(x: f32, y: f32, r: f32, color: i32);
-    pub fn circle_fill(x: f32, y: f32, r: f32, color: i32);
+    pub fn circle(x: i32, y: i32, r: i32, color: i32);
+    pub fn circle_fill(x: i32, y: i32, r: i32, color: i32);
     pub fn print(ptr: *const u8, len: u32, x: i32, y: i32, color: i32) -> i32;
     pub fn is_button_down(b: u32) -> i32;
     pub fn is_button_pressed(b: u32) -> i32;
@@ -88,8 +88,8 @@ mod stubs {
     pub unsafe fn line(_x0: i32, _y0: i32, _x1: i32, _y1: i32, _color: i32) {}
     pub unsafe fn rect(_x0: f32, _y0: f32, _x1: f32, _y1: f32, _color: i32) {}
     pub unsafe fn rect_fill(_x0: f32, _y0: f32, _x1: f32, _y1: f32, _color: i32) {}
-    pub unsafe fn circle(_x: f32, _y: f32, _r: f32, _color: i32) {}
-    pub unsafe fn circle_fill(_x: f32, _y: f32, _r: f32, _color: i32) {}
+    pub unsafe fn circle(_x: i32, _y: i32, _r: i32, _color: i32) {}
+    pub unsafe fn circle_fill(_x: i32, _y: i32, _r: i32, _color: i32) {}
     pub unsafe fn print(_ptr: *const u8, _len: u32, _x: i32, _y: i32, _color: i32) -> i32 {
         0
     }
